Создание общей библиотеки, создание DTO, а также Pipe соединения, модификация модулей и добавление нового интерфейса, добавление новых модулей, соединение службы и GUI

+++ Contracts/AppConfigDto.cs
@@ -3,7 +3,7 @@ namespace Contracts
     public sealed class AppConfigDto
     {
         public string[] InnList { get; set; } = [];
-        public int IntervalHours { get; set; } = 24;
+        public double IntervalHours { get; set; } = 24;
         public bool Test { get; set; }
         public string FilePathToSavedHtml { get; set; } = "";
         public string FilePathToLogs { get; set; } = "";
